feat: Refatored ClosedChats separating RoomsTable in a compontent
